% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_cols_group.R
\name{avg_cols_group}
\alias{avg_cols_group}
\title{Column Averages Per Group}
\usage{
avg_cols_group(.data, y, na_rm = TRUE)
}
\arguments{
\item{.data}{The data frame or data frame extension (e.g. tibble). We opted for the function name ".data" to avoid name collision with the \code{data()} function in base R packages.}

\item{y}{The variable to group by. This is interpreted as a categorical variable. If left unspecified, the function will still work, though will simply be computing means of all rows since no grouping parameter was specified.}

\item{na_rm}{A logical value (\code{TRUE} / \code{FALSE}) indicating whether \code{NA} values should be stripped before the computation proceeds. If \code{TRUE} (the default), the calculation will first remove all \code{NA}s (if any element of \code{y} or another grouping variable is \code{NA}), and the result should therefore not contain \code{NA}s.}
}
\value{
Returns a data frame containing to the average value of each numeric column per level of the grouping variable (if there is one, or more). With the exception of column names as the top row of the dataframe, each subsequent row of the dataframe corresponds to a grouping 'level', and contains the means -- for that level -- of all the numeric variables in the dataset.
}
\description{
This function first identifies a grouping variable (taken as categorical). Inspecting across all columns of the data frame, this function then selects all of the numeric columns and calculates the means of each of those numeric columns for each group (i.e., level of the grouping variable, corresponding to particular rows).
}
\examples{
avg_cols_group(.data = mtcars, y = cyl)
    # Should output a 3 x 11 tibble with one row corresponding to each `cyl` group
    # (4, 6, or 8 cylinder cars) as well as their mean values for each variable

avg_cols_group(.data = palmerpenguins::penguins, y = species)
    # Should output a 3 x 6 tibble with one row corresponding to each penguin species,
    # and each non-`species` column a numeric variable;
    # numbers in the output tibble should each correspond to the average of a specific numeric variable
    # for a specific species (e.g., average bill_length_mm of Adelie penguins = 38.8)
}
